requirements.txt
langchain
langchain-openai
langchain-ollama
langchain_community
chromadb
tiktoken
openai
cohere
python-dotenv
typer
pytest
numpy
requests
tenacity
beautifulsoup4
sentence-transformers
transformers
InstructorEmbedding
accelerate
torch
pypdf
unstructured
pdfminer.six
